requirements.txt
Django>=4.2
djangorestframework
gunicorn
psycopg2-binary
django-environ
celery==5.3.1
django-celery-beat==2.5.0
flower==2.0.1
channels==4.0.0


